+++ include/nstd/audio.h
@@ -1,9 +1,9 @@
 #ifndef NSTD_AUDIO_H_INCLUDED
 #define NSTD_AUDIO_H_INCLUDED
-#include "audio/audio.h"
 #include "audio/def.h"
 #include "audio/device.h"
 #include "audio/host.h"
+#include "audio/player.h"
 #include "audio/sink.h"
 #include "audio/stream.h"
 #include "nstd.h"