+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - Moved `nstd_sys` contents into `nstd_env`, removing `nstd_sys` as a result.
 ### `nstd.audio`
 - Renamed `NSTDAudioPlayStream` to `NSTDAudioPlayer`.
-- Added `def`, `device`, `host`, `sink`, and `stream` modules.
+- Added `def`, `device`, `host`, `player`, `sink`, and `stream` modules.
 ### `nstd.core`
 - Removed `nstd_core_abort` & `nstd_core_assert`.
 ### `nstd.events`